fix: in partners remove each components and create map functiontr5
