--- src/components/segments/areas/industry.jsx
@@ -1,17 +1,0 @@
-import Title from "../title";
-
-function Industry(){
-  return(
-    <>
-      <section className='flex flex-col space-y-4'>
-        <Title
-          title='Indústria'
-          icon='iconoir:industry'
-        />
-        <p>Morbi sem diam, consequat id ullamcorper et, dignissim ut lorem. Sed euismod diam eu leo feugiat, non accumsan metus iaculis. Nulla at gravida ex. Nunc non tincidunt neque, eget ornare leo. Class aptent taciti sociosqu ad litora torquent per conubia.</p>
-      </section>
-    </>
-  )
-}
-
-export default Industry;
