--- src/components/segments/areas/network.jsx
@@ -1,17 +1,0 @@
-import Title from "../title";
-
-function Network(){
-  return(
-    <>
-      <section className='flex flex-col space-y-4'>
-        <Title
-          title='Redes'
-          icon='iconoir:www'
-        />
-        <p>Phasellus iaculis metus ut turpis viverra pellentesque. Nunc ultricies sodales arcu non hendrerit. Maecenas quis nulla sit amet nibh tempus consectetur. Pellentesque habitant morbi tristique senectus et netus et malesuada fames ac turpis egestas. In eu ante non erat lobortis.</p>
-      </section>
-    </>
-  )
-}
-
-export default Network;
